Fix issues with type checking and deployment preparation
Update type definitions and casting for better TypeScript compatibility, and adjust the pre-push script to correctly build shared library declarations.

+++ artifacts/que-lo-que/src/hooks/usePushNotifications.ts
@@ -73,7 +73,7 @@ export function usePushNotifications() {
       const reg = await navigator.serviceWorker.ready;
       const sub = await reg.pushManager.subscribe({
         userVisibleOnly: true,
-        applicationServerKey: urlBase64ToUint8Array(publicKey),
+        applicationServerKey: urlBase64ToUint8Array(publicKey) as Uint8Array,
       });
       await subscribeOnServer(sub);
       setSubscription(sub);
